requirements.txt
streamlit
pandas
google-generativeai>=0.3.0
python-dotenv